pyproject.toml
[project]
name = "chimpanzukiaiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fpdf2>=2.8.3",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.7",
    "langchain-tavily>=0.2.7",
    "langgraph>=0.5.2",
    "pyfiglet>=1.0.3",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "pytest>=8.4.1",
    "rapidocr-onnxruntime>=1.4.4",
    "tavily-python>=0.7.9",
]
